feat: harden syscall and networking foundations
- document the INT 0x80 syscall ABI and supported operations
- add controlled syscall shell access with rex protection
- add real wifi-list scanning command
- remove fake Wi-Fi devices, fallback SSIDs, and fake USB success paths
- preserve 64-bit PCI BAR addresses
- use coherent DMA allocation for e1000 rings and buffers
- add e1000 IRQ registration and interrupt acknowledgement
- implement bounded e1000 RX/TX ring ownership and completion tracking
- add MMIO memory barriers and TX ring backpressure
- enable QEMU user-mode NAT networking in all launch scripts
- fix 64-bit QEMU selection in simp_run.sh
- build mkiofs natively for the host environment
- document networking limitations and verification requirements

Validation:
- make clean && make -j2
- bash -n scripts/run.sh scripts/run_fullscreen.sh scripts/simp_run.sh
- git diff --check

Networking remains intentionally incomplete for socket syscalls, DHCP,
DNS, TCP listen/accept, USB Wi-Fi, and nmap until their real subsystems
are implemented.\

+++ include/pci.h
@@ -39,7 +39,7 @@ typedef struct pci_device {
     u8 subclass;
     u8 prog_if;
     u8 header_type;
-    u32 bars[6];
+    u64 bars[6];
     u8 bar_is_mem[6];
     u8 irq_line;
     struct pci_device *next;
